requirement.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-js-asset==2.0.0
django-mptt==0.13.4
factory-boy==3.2.1
Faker==13.14.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-dateutil==2.8.2
requests==2.28.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
